server/requirements.txt
mysql-connector-python>=8.0
matplotlib>=3.0
pandas>=1.0
